ACCUMULO-4820 More code cleanup for 2.0 (#398)
* Replace explicit types with <>
* Use try with resources where possible
* Remove unnecessary boxing & unboxing
* Drop unnecessary boolean expressions
* Use Java 1.5 foreach where possible
* Remove redundant type arguments & casting

+++ server/base/src/main/java/org/apache/accumulo/server/master/balancer/ChaoticLoadBalancer.java
@@ -89,7 +89,7 @@ public class ChaoticLoadBalancer extends TabletBalancer {
       int index = r.nextInt(tServerArray.size());
       TServerInstance dest = tServerArray.get(index);
       assignments.put(ke, dest);
-      long remaining = toAssign.get(dest).longValue() - 1;
+      long remaining = toAssign.get(dest) - 1;
       if (remaining == 0) {
         tServerArray.remove(index);
         toAssign.remove(dest);
